# wwwroot/swagger/v1/swagger.json
{
...
  "openapi": "3.0.0",
  "info": {
    "title": "Api Comercial",
    "description": "API para la gestión de documentos comerciales",
    "contact": {
      "name": "Soporte API",
      "email": ""
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Ventas/VentasContado": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "summary": "Consulta las ventas al contado",
        "operationId": "Ventas_VentasContado",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/VentasResponse"
                  }
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ventas/MediosPago": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "operationId": "Ventas_MediosPago",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/MediosPagoResponse"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ventas/cabeceraCuota": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "operationId": "Ventas_CabeceraCuota",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CabeceraCuotaResponse"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ventas/{idVenta}/DetalleCuota": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "operationId": "Ventas_DetalleCuota",
        "parameters": [
          {
            "name": "idVenta",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DetalleCuotaResponse"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ventas/PagarCuota": {
      "put": {
        "tags": [
          "Ventas"
        ],
        "operationId": "Ventas_PagarCuota",
        "requestBody": {
          "x-name": "cuota",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PagarCuotaRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PagarCuotaRequest"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ventas/{idVenta}/Refuerzos": {
      "get": {
        "tags": [
          "Ventas"
        ],
        "operationId": "Ventas_ObtenerRefuerzos",
        "parameters": [
          {
            "name": "idVenta",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefuerzoResponse"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ventas/InsertarRefuerzo": {
      "post": {
        "tags": [
          "Ventas"
        ],
        "operationId": "Ventas_InsertarRefuerzo",
        "requestBody": {
          "x-name": "parametros",
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/RefuerzoRequest"
              }
            }
          },
          "required": true,
          "x-position": 1
        },
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RefuerzoResponse"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/Ventas/EliminarVentaCuotas/{idVenta}": {
      "delete": {
        "tags": [
          "Ventas"
        ],
        "operationId": "Ventas_EliminarVentaCuotas",
        "parameters": [
          {
            "name": "idVenta",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int32"
            },
            "x-position": 1
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "boolean"
                }
              }
            }
          },
          "204": {
            "description": ""
          },
          "400": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "500": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "ErrorResponse": {
        "type": "object",
        "description": "Clase que muestra los errores",
        "additionalProperties": false,
        "properties": {
          "errorType": {
            "$ref": "#/components/schemas/ErrorType"
          },
          "errorDescripcion": {
            "type": "string",
            "nullable": true
          }
        }
      },
      "ErrorType": {
        "type": "integer",
        "description": "",
        "x-enumNames": [
          "error_interno_servidor",
          "validacion_negocio"
        ],
        "enum": [
          0,
          1
        ]
      },
...
      "VentasResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "ventaId": {
            "type": "integer",
            "format": "int32"
          },
          "idCliente": {
            "type": "integer",
            "format": "int32"
          },
          "cedulaCliente": {
            "type": "string",
            "nullable": true
          },
          "nombreCliente": {
            "type": "string",
            "nullable": true
          },
          "idChasis": {
            "type": "string",
            "nullable": true
          },
          "marca": {
            "type": "string",
            "nullable": true
          },
          "anoFabricacion": {
            "type": "integer",
            "format": "int32"
          },
          "precio": {
            "type": "string",
            "nullable": true
          },
          "fechaCompra": {
            "type": "string",
            "format": "date-time"
          }
        }
      },
      "MediosPagoResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "medioPagoId": {
            "type": "integer",
            "description": "Id del medio de pago",
            "format": "int32"
          },
          "codigo": {
            "type": "string",
            "description": "TC, TD, CHEQUE, EFECTIVO, TRANSFER"
          },
          "nombre": {
            "type": "string",
            "description": "Descripción del medio de pago",
            "nullable": true
          }
        }
      },
      "CabeceraCuotaResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "idVenta": {
            "type": "integer",
            "format": "int32"
          },
          "cedulaCliente": {
            "type": "string",
            "nullable": true
          },
          "nombreCliente": {
            "type": "string",
            "nullable": true
          },
          "idChasis": {
            "type": "string",
            "nullable": true
          },
          "marca": {
            "type": "string",
            "nullable": true
          },
          "anoFabricacion": {
            "type": "integer",
            "format": "int32"
          },
          "precio": {
            "type": "string",
            "nullable": true
          },
          "interes": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "cantidadCuotas": {
            "type": "integer",
            "format": "int32"
          },
          "precioTotal": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "fechaVenta": {
            "type": "string",
            "format": "date-time"
          },
          "cantidadCuotasPagadas": {
            "type": "integer",
            "format": "int32"
          },
          "totalPagado": {
            "type": "integer",
            "format": "int32"
          },
          "totalRestante": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          }
        }
      },
      "DetalleCuotaResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "ventaId": {
            "type": "integer",
            "format": "int32"
          },
          "idCuota": {
            "type": "integer",
            "format": "int32"
          },
          "numeroCuota": {
            "type": "integer",
            "format": "int32"
          },
          "montoCuota": {
            "type": "integer",
            "format": "int32"
          },
          "fechaVencimiento": {
            "type": "string",
            "format": "date-time"
          },
          "estadoCodigo": {
            "type": "string",
            "nullable": true
          },
          "esRefuerzo": {
            "type": "boolean"
          },
          "idRefuerzo": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "montoRefuerzo": {
            "type": "number",
            "format": "decimal",
            "nullable": true
          },
          "fechaVencimientoRefuerzo": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          }
        }
      },
      "PagarCuotaRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "cuotaId": {
            "type": "integer",
            "description": "Id del medio de pago",
            "format": "int32"
          },
          "medioPagoId": {
            "type": "integer",
            "description": "Monto a pagar",
            "format": "int32"
          },
          "montoPagado": {
            "type": "integer",
            "description": "Fecha de pago",
            "format": "int32"
          },
          "referencia": {
            "type": "string",
            "description": "Observaciones del pago",
            "nullable": true
          }
        }
      },
      "RefuerzoResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "refuerzoId": {
            "type": "integer",
            "format": "int32"
          },
          "ventaId": {
            "type": "integer",
            "description": "Id de la cuota a la que se le aplica el refuerzo",
            "format": "int32"
          },
          "montoRefuerzo": {
            "type": "number",
            "description": "Monto del refuerzo",
            "format": "decimal"
          },
          "fechaVencimiento": {
            "type": "string",
            "description": "Fecha de vencimiento del refuerzo",
            "format": "date-time"
          }
        }
      },
      "RefuerzoRequest": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "ventaId": {
            "type": "integer",
            "format": "int32"
          },
          "montoRefuerzo": {
            "type": "number",
            "format": "decimal"
          },
          "fechaVencimiento": {
            "type": "string",
            "format": "date-time"
          }
        }
      }
    },
    "securitySchemes": {
      "JWT": {
        "type": "apiKey",
        "description": "Ingrese: Bearer {su token JWT}",
        "name": "Authorization",
        "in": "header"
      }
    }
  },
  "security": [
    {
      "JWT": []
    }
  ]
}